Добавил модуль print-photo
