+++ js/create-photo-album.js
@@ -24,35 +24,7 @@ const createPostPhoto = () => {
   };
 };
 
+// Создаем функуию, которая возвращает массив объектов (фото+комментарии и т.п.)
 const createPhotoAlbum = () => Array.from({length:NUMBER_ID_PHOTOS}, createPostPhoto);
 
-
-const pictureTemplate = document.querySelector('#picture').content.querySelector('a');
-const photoAlbum = createPhotoAlbum();
-const photoAlbumFragment = document.createDocumentFragment();
-const picturesContainer = document.querySelector('.pictures');
-
-photoAlbum.forEach((element) => {
-  const newPicture = pictureTemplate.cloneNode(true);
-  newPicture.querySelector('.picture__img').src = element.url;
-  newPicture.querySelector('.picture__likes').textContent = element.likes;
-  newPicture.querySelector('.picture__comments').textContent = element.comments.length;
-  photoAlbumFragment.append(newPicture);
-});
-
-picturesContainer.append(photoAlbumFragment);
-
 export {createPhotoAlbum};
-
-
-// Заведите модуль, который будет отвечать за отрисовку миниатюр.
-
-// На основе временных данных для разработки и шаблона #picture создайте DOM-элементы, соответствующие фотографиям,
-//  и заполните их данными:
-
-// Адрес изображения url подставьте как атрибут src изображения.
-// Количество лайков likes выведите в блок .picture__likes.
-// Количество комментариев comments выведите в блок .picture__comments.
-// Отрисуйте сгенерированные DOM-элементы в блок .pictures. Для вставки элементов используйте DocumentFragment.
-
-// Подключите модуль в проект.
